improvement(ui/ux): logs control bar
